Code Cleanup
- Removed redundant code.
